+++ inst/examples/supported_shinyWidgets_app/ui.R
@@ -6,7 +6,6 @@ fluidPage(
     column(
       4,
       h2("shinyWidgets::airDatepickerInput"),
-      
       shinyWidgets::airDatepickerInput(
         inputId = "myAirDatePickerInput",
         label = "Danger if not today"
@@ -20,6 +19,16 @@ fluidPage(
         label = "Danger if A",
         choices = c("A", "B"),
         selected = "A"
+      )
+    ),
+    column(
+      4,
+      h2("shinyWidgets::currencyInput"),
+      shinyWidgets::currencyInput(
+        inputId = "myCurrencyInput",
+        label = "Danger if 0 or less (or empty)",
+        format = "dollar",
+        value = 100
       )
     )
   )